requirements.txt
fastapi[standard]>=0.100.0
pyyaml>=6.0
uvicorn>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
requests>=2.31.0
streamlit>=1.20.0
plotly>=5.10.0
pandas>=1.5.0
